+++ documentation/changelog.md
@@ -82,6 +82,10 @@ please use _ni_x_series_in_streamer.py_ as hardware module.
 * Set proper minimum wavelength value in constraints of Tektronix AWG7k series HW module
 * Added a hardware file for fibered optical switch Thorlabs OSW12/22 via SwitchInterface
 * Fixed bug affecting interface overloading of Qudi modules
+<<<<<<< HEAD
+* Added a hardware file to interface Thorlabs MFF101 flipper via SwitchInterface
+*
+=======
 * Added ScienDSpinbox and ScienSpinbox compatibility to mapper
 * Added missing metadata in saved raw data file of PulsedMeasurement module
 * Update hardware module controlling the cryocon temperature regulator
@@ -95,6 +99,7 @@ please use _ni_x_series_in_streamer.py_ as hardware module.
 * Added possibility to fit data of all ranges in ODMR module when Fit range is -1
 * Added new dummy/interface/logic/gui chain for scientific spectrometer
 * Added new spectrometer and camera hardware modules for andor devices
+>>>>>>> spectrometer_PR_v1
 
 
 Config changes: